3.2测试若干bug 修改

+++ app/src/main/java/com/cylan/jiafeigou/cache/db/impl/BaseDBHelper.java
@@ -461,7 +461,9 @@ public class BaseDBHelper implements IDBHelper {
     public Observable<DPEntity> deleteDPMsgForce(String account, String server, String uuid, Long version, Integer msgId) {
         return applyUnique(buildDPMsgQueryBuilder(account, server, uuid, version, msgId, null, null, null)
                 .rx()).map(result -> {
-            mEntityDao.delete(result);
+            if (result != null) {
+                mEntityDao.delete(result);
+            }
             return result;
         });
     }
